lemmatizer/requirements.txt
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cymem==2.0.6
DAWG-Python==0.7.2
docopt==0.6.2
grpcio==1.47.0
grpcio-tools==1.47.0
idna==3.3
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.8
numpy==1.23.2
packaging==21.3
pathy==0.6.2
preshed==3.0.7
protobuf==3.20.1
pydantic==1.9.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.1
ru-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ru_core_news_sm-3.4.0/ru_core_news_sm-3.4.0-py3-none-any.whl
six==1.16.0
smart-open==5.2.1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
thinc==8.1.0
tqdm==4.64.0
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
